move KV ICQ cosmos-sdk v0.45 specific stuff to a dedicated package
