+++ packages/neutron-sdk/src/interchain_queries/mod.rs
@@ -1,11 +1,8 @@
 pub mod helpers;
 pub mod queries;
-pub mod register_queries;
 pub mod types;
+pub mod v045;
 
-pub use register_queries::{
-    new_register_balance_query_msg, new_register_bank_total_supply_query_msg,
-    new_register_delegator_delegations_query_msg, new_register_distribution_fee_pool_query_msg,
-    new_register_gov_proposal_query_msg, new_register_staking_validators_query_msg,
-    new_register_transfers_query_msg,
+pub use queries::{
+    check_query_type, get_registered_query, new_register_interchain_query_msg, query_kv_result,
 };
