+++ packages/neutron-sdk/src/interchain_queries/queries.rs
@@ -2,56 +2,14 @@ use crate::bindings::query::{
     InterchainQueries, QueryRegisteredQueryResponse, QueryRegisteredQueryResultResponse,
 };
 use crate::interchain_queries::types::{
-    Balances, Delegations, FeePool, GovernmentProposal, KVReconstruct, QueryType, StakingValidator,
-    TotalSupply,
+    KVReconstruct, QueryPayload, QueryType, TransactionFilterItem,
 };
-
-use crate::{NeutronError, NeutronResult};
-use cosmwasm_std::{Deps, Env};
-use schemars::JsonSchema;
-use serde::{Deserialize, Serialize};
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub struct BalanceResponse {
-    pub balances: Balances,
-    pub last_submitted_local_height: u64,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub struct TotalSupplyResponse {
-    pub supply: TotalSupply,
-    pub last_submitted_local_height: u64,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub struct FeePoolResponse {
-    pub pool: FeePool,
-    pub last_submitted_local_height: u64,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub struct ValidatorResponse {
-    pub validator: StakingValidator,
-    pub last_submitted_local_height: u64,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub struct ProposalResponse {
-    pub proposals: GovernmentProposal,
-    pub last_submitted_local_height: u64,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub struct DelegatorDelegationsResponse {
-    pub delegations: Vec<cosmwasm_std::Delegation>,
-    pub last_submitted_local_height: u64,
-}
+use crate::NeutronError;
+use crate::{
+    bindings::{msg::NeutronMsg, types::KVKey},
+    errors::error::NeutronResult,
+};
+use cosmwasm_std::{Deps, DepsMut, Env};
 
 /// Checks **actual** query type is **expected** query type
 pub fn check_query_type(actual: QueryType, expected: QueryType) -> NeutronResult<()> {
@@ -61,6 +19,31 @@ pub fn check_query_type(actual: QueryType, expected: QueryType) -> NeutronResult
         });
     }
     Ok(())
+}
+
+#[allow(clippy::too_many_arguments)]
+/// Creates a message to register an Interchain Query with provided params
+pub fn new_register_interchain_query_msg(
+    _deps: DepsMut<InterchainQueries>,
+    _env: Env,
+    connection_id: String,
+    query_type: QueryType,
+    kv_keys: Vec<KVKey>,
+    transactions_filter: Vec<TransactionFilterItem>,
+    update_period: u64,
+) -> NeutronResult<NeutronMsg> {
+    match query_type {
+        QueryType::KV => NeutronMsg::register_interchain_query(
+            QueryPayload::KV(kv_keys),
+            connection_id,
+            update_period,
+        ),
+        QueryType::TX => NeutronMsg::register_interchain_query(
+            QueryPayload::TX(transactions_filter),
+            connection_id,
+            update_period,
+        ),
+    }
 }
 
 /// Queries registered query info
@@ -76,18 +59,6 @@ pub fn get_registered_query(
     Ok(res)
 }
 
-/// Queries interchain query result (raw KV storage values or transactions) from Interchain Queries Module
-fn get_interchain_query_result(
-    deps: Deps<InterchainQueries>,
-    interchain_query_id: u64,
-) -> NeutronResult<QueryRegisteredQueryResultResponse> {
-    let interchain_query = InterchainQueries::InterchainQueryResult {
-        query_id: interchain_query_id,
-    };
-    let res = deps.querier.query(&interchain_query.into())?;
-    Ok(res)
-}
-
 /// Reads submitted raw KV values for Interchain Query with **query_id** from the storage and reconstructs the result
 pub fn query_kv_result<T: KVReconstruct>(
     deps: Deps<InterchainQueries>,
@@ -98,128 +69,14 @@ pub fn query_kv_result<T: KVReconstruct>(
     KVReconstruct::reconstruct(&registered_query_result.result.kv_results)
 }
 
-/// Returns balance of account on remote chain for particular denom
-/// * ***registered_query_id*** is an identifier of the corresponding registered interchain query
-pub fn query_balance(
+/// Queries interchain query result (raw KV storage values or transactions) from Interchain Queries Module
+fn get_interchain_query_result(
     deps: Deps<InterchainQueries>,
-    _env: Env,
-    registered_query_id: u64,
-) -> NeutronResult<BalanceResponse> {
-    let registered_query = get_registered_query(deps, registered_query_id)?;
-
-    check_query_type(registered_query.registered_query.query_type, QueryType::KV)?;
-
-    let balances: Balances = query_kv_result(deps, registered_query_id)?;
-
-    Ok(BalanceResponse {
-        last_submitted_local_height: registered_query
-            .registered_query
-            .last_submitted_result_local_height,
-        balances,
-    })
-}
-
-/// Returns bank total supply on remote chain for particular denom
-/// * ***registered_query_id*** is an identifier of the corresponding registered interchain query
-pub fn query_bank_total(
-    deps: Deps<InterchainQueries>,
-    _env: Env,
-    registered_query_id: u64,
-) -> NeutronResult<TotalSupplyResponse> {
-    let registered_query = get_registered_query(deps, registered_query_id)?;
-
-    check_query_type(registered_query.registered_query.query_type, QueryType::KV)?;
-
-    let total_supply: TotalSupply = query_kv_result(deps, registered_query_id)?;
-
-    Ok(TotalSupplyResponse {
-        last_submitted_local_height: registered_query
-            .registered_query
-            .last_submitted_result_local_height,
-        supply: total_supply,
-    })
-}
-
-/// Returns distribution fee pool on remote chain
-/// * ***registered_query_id*** is an identifier of the corresponding registered interchain query
-pub fn query_distribution_fee_pool(
-    deps: Deps<InterchainQueries>,
-    _env: Env,
-    registered_query_id: u64,
-) -> NeutronResult<FeePoolResponse> {
-    let registered_query = get_registered_query(deps, registered_query_id)?;
-
-    check_query_type(registered_query.registered_query.query_type, QueryType::KV)?;
-
-    let fee_pool: FeePool = query_kv_result(deps, registered_query_id)?;
-
-    Ok(FeePoolResponse {
-        last_submitted_local_height: registered_query
-            .registered_query
-            .last_submitted_result_local_height,
-        pool: fee_pool,
-    })
-}
-
-/// Returns staking validator from remote chain
-/// * ***registered_query_id*** is an identifier of the corresponding registered interchain query
-pub fn query_staking_validators(
-    deps: Deps<InterchainQueries>,
-    _env: Env,
-    registered_query_id: u64,
-) -> NeutronResult<ValidatorResponse> {
-    let registered_query = get_registered_query(deps, registered_query_id)?;
-
-    check_query_type(registered_query.registered_query.query_type, QueryType::KV)?;
-
-    let validator: StakingValidator = query_kv_result(deps, registered_query_id)?;
-
-    Ok(ValidatorResponse {
-        last_submitted_local_height: registered_query
-            .registered_query
-            .last_submitted_result_local_height,
-        validator,
-    })
-}
-
-/// Returns list of government proposals on the remote chain
-/// * ***registered_query_id*** is an identifier of the corresponding registered interchain query
-pub fn query_government_proposals(
-    deps: Deps<InterchainQueries>,
-    _env: Env,
-    registered_query_id: u64,
-) -> NeutronResult<ProposalResponse> {
-    let registered_query = get_registered_query(deps, registered_query_id)?;
-
-    check_query_type(registered_query.registered_query.query_type, QueryType::KV)?;
-
-    let proposals: GovernmentProposal = query_kv_result(deps, registered_query_id)?;
-
-    Ok(ProposalResponse {
-        last_submitted_local_height: registered_query
-            .registered_query
-            .last_submitted_result_local_height,
-        proposals,
-    })
-}
-
-/// Returns delegations of particular delegator on remote chain
-/// * ***registered_query_id*** is an identifier of the corresponding registered interchain query
-pub fn query_delegations(
-    deps: Deps<InterchainQueries>,
-    _env: Env,
-    registered_query_id: u64,
-) -> NeutronResult<DelegatorDelegationsResponse> {
-    let registered_query = get_registered_query(deps, registered_query_id)?;
-
-    check_query_type(registered_query.registered_query.query_type, QueryType::KV)?;
-
-    let delegations: Delegations = query_kv_result(deps, registered_query_id)?;
-
-    Ok(DelegatorDelegationsResponse {
-        delegations: delegations.delegations,
-        last_submitted_local_height: registered_query
-            .registered_query
-            .last_submitted_result_local_height,
-    })
+    interchain_query_id: u64,
+) -> NeutronResult<QueryRegisteredQueryResultResponse> {
+    let interchain_query = InterchainQueries::InterchainQueryResult {
+        query_id: interchain_query_id,
+    };
+    let res = deps.querier.query(&interchain_query.into())?;
+    Ok(res)
 }
